add more primitives

+++ crates/vm_library/src/ticket_table.rs
@@ -21,7 +21,7 @@ pub enum Error {
 pub type Result<T> = std::result::Result<T, Error>;
 
 type Address = String;
-type Amount = u32;
+type Amount = usize;
 type Handle = usize;
 
 fn assert_not_dead(ticket: &Ticket) -> Result<()> {
